debundle/lowering: extract materialize
BAZEL_TEST_INVOCATIONS=none: file move.
